Fix local GQL parser allowing invalid syntax
Fix on the 'WHERE' clause some invalid syntax might be allowed on the
local GQL parser. This will check if the property exists and if the
right hand side is properly represented as its type. This fixes when the
string not being properly quoted might run successfully on the local GQL
parser.